feat : new map more accurate fix : overflow in flag id

+++ Source/exploration_patrol/Public/Flags/SO_Flag.h
@@ -13,7 +13,7 @@ struct FFlagSegment
 {
 	GENERATED_BODY()
 	UPROPERTY(EditAnywhere)
-	uint8 id;
+	int id;
 	UPROPERTY(EditAnywhere)
 	FVector BeginPosition;
 	UPROPERTY(EditAnywhere)
